potado/requirements.txt
black
coverage
isort
mock
pylama
pylint
pytest
pytest-cov
pytest-random-order
requests
requests-mock
ruamel.yaml

ipdb
